+++ MVCoin/SatellitesControl.cs
@@ -17,18 +17,19 @@ namespace MVCoin
         List<Satellite> satelliteList = new List<Satellite>();
         List<Animation> satelliteAniList = new List<Animation>();
         List<Point> satelliteOffsetList = new List<Point>();
-        List<taskName> taskList = new List<taskName>() {taskName.STICKIES };
+        List<taskName> taskList = new List<taskName>() {taskName.STICKIES, taskName.YTVIEWER};
         double scaleFactor = 1;
-        int satelliteNumber = 6;
+        int satelliteNumber;
         int radiusOrbit = 100;
         int i = 0; // Counter
 
         string[] imgPath =
-            {"../../Icon/sticky.png"};
+            {"../../Icon/sticky.png", "../../Icon/youtube.png"};
 
         public SatellitesControl(Form mainFormInput)
         {
             mainForm = mainFormInput;
+            satelliteNumber = taskList.Count();
         }
 
         public void setSatelliteNum(int numberInput)