+++ MVCoin/YTViewer.Designer.cs
@@ -1,0 +1,80 @@
+﻿namespace MVCoin
+{
+    partial class YTViewer
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.webBrowser = new System.Windows.Forms.WebBrowser();
+            this.URLtextBox = new System.Windows.Forms.TextBox();
+            this.SuspendLayout();
+            // 
+            // webBrowser
+            // 
+            this.webBrowser.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.webBrowser.Location = new System.Drawing.Point(0, 0);
+            this.webBrowser.MinimumSize = new System.Drawing.Size(20, 20);
+            this.webBrowser.Name = "webBrowser";
+            this.webBrowser.ScrollBarsEnabled = false;
+            this.webBrowser.Size = new System.Drawing.Size(578, 419);
+            this.webBrowser.TabIndex = 3;
+            // 
+            // URLtextBox
+            // 
+            this.URLtextBox.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.URLtextBox.Font = new System.Drawing.Font("Berlin Sans FB", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.URLtextBox.ForeColor = System.Drawing.Color.Red;
+            this.URLtextBox.Location = new System.Drawing.Point(12, 12);
+            this.URLtextBox.Name = "URLtextBox";
+            this.URLtextBox.Size = new System.Drawing.Size(554, 25);
+            this.URLtextBox.TabIndex = 4;
+            this.URLtextBox.Text = "Paste your youtube URL";
+            this.URLtextBox.TextChanged += new System.EventHandler(this.URLtextBox_TextChanged);
+            // 
+            // YTViewer
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 12F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(578, 419);
+            this.Controls.Add(this.URLtextBox);
+            this.Controls.Add(this.webBrowser);
+            this.DoubleBuffered = true;
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            this.Name = "YTViewer";
+            this.ShowInTaskbar = false;
+            this.Text = "YTViewer";
+            this.TopMost = true;
+            this.Load += new System.EventHandler(this.YTViewer_Load);
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+        private System.Windows.Forms.WebBrowser webBrowser;
+        private System.Windows.Forms.TextBox URLtextBox;
+    }
+}